requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
graphviz==0.8.4
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
imutils==0.5.4
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
mxnet
networkx==3.2.1
numpy==1.23.5
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.126
ultralytics-thop==2.0.14
urllib3==2.4.0
zipp==3.21.0
